% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockr.input-package.R
\docType{package}
\name{blockr.input-package}
\alias{blockr.input}
\alias{blockr.input-package}
\title{blockr.input: Interactive data-entry blocks for blockr}
\description{
Extends 'blockr.core' with blocks that let an end user edit data inside a
running dashboard: \code{\link[=new_form_edit_block]{new_form_edit_block()}} for
row-level edits, \code{\link[=new_grid_edit_block]{new_grid_edit_block()}} for spreadsheet-style bulk entry,
and \code{\link[=new_table_edit_block]{new_table_edit_block()}} for server-paginated browse-and-edit. Each
emits ordinary \code{dplyr::rows_upsert()} / \code{dplyr::rows_delete()} steps.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BristolMyersSquibb/blockr.input}
  \item Report bugs at \url{https://github.com/BristolMyersSquibb/blockr.input/issues}
}

}
\author{
\strong{Maintainer}: Christoph Sax \email{christoph@cynkra.com} (\href{https://orcid.org/0000-0002-7192-7044}{ORCID})

}
\keyword{internal}
