% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit-block-dep.R
\name{edit_block_dep}
\alias{edit_block_dep}
\title{HTML dependency for the edit block}
\usage{
edit_block_dep()
}
\value{
An \code{htmltools::tagList} of \code{htmlDependency} objects.
}
\description{
Bundles the JS class + CSS for \code{new_edit_block()}. Exported so other blockr
packages can embed the edit-block UI.
}
