% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_block.R
\name{new_table_block}
\alias{new_table_block}
\title{Table block (deprecated alias)}
\usage{
new_table_block(...)
}
\arguments{
\item{...}{Forwarded to \code{\link[=new_table_crud_block]{new_table_crud_block()}}.}
}
\value{
A blockr block of class \code{table_crud_block}.
}
\description{
Deprecated alias for \code{\link[=new_table_crud_block]{new_table_crud_block()}}. The block was renamed
from \code{new_table_block} to \code{new_table_crud_block} to free the generic
"Table" name. Kept so existing serialized boards that reference
\code{new_table_block} still deserialize.
}
